requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
boto3==1.28.78
botocore==1.31.78
cent==4.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Faker==20.0.3
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
itsdangerous==2.1.2
jmespath==1.0.1
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pymemcache==4.0.0
pypika-tortoise==0.1.6
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.20.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
watchfiles==0.21.0
websockets==12.0
